Remove typescript
