--- src/UserAware.tsx
@@ -1,24 +1,0 @@
-import { Component } from "react";
-import { connect } from "react-redux";
-import { preloadUser } from "./actions/typedActionCreators";
-
-class UserAware extends Component<{ preload: () => void }> {
-  public componentWillMount() {
-    this.props.preload();
-  }
-
-  public render() {
-    return this.props.children;
-  }
-}
-
-function mapDispatchToProps(dispatch: Function) {
-  return {
-    preload: () => dispatch(preloadUser())
-  };
-}
-
-export default connect(
-  null,
-  mapDispatchToProps
-)(UserAware);
