+++ src/index.jsx
@@ -1,9 +1,8 @@
 import * as React from 'react';
-
+import { Provider } from 'react-redux';
 import { render } from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import configureStore from './store/configureStore';
-import {Provider} from 'react-redux';
 import App from './App';
 import './index.css';
 
@@ -11,7 +10,7 @@ const store = configureStore();
 
 render(
   <Provider store={store}>
-      <App />
+    <App />
   </Provider>,
   document.getElementById('root')
 );
