+++ src/reducers/commandOutput.js
@@ -1,15 +1,6 @@
-import { EXECUTE_COMMAND_COMPLETE } from "../actions/actionTypes";
+import { EXECUTE_COMMAND_COMPLETE } from '../actions/actionTypes';
 
-export interface ICommandOutputState {
-  [device: string]: {
-    [command: string]: any;
-  };
-}
-
-export default function commandOutput(
-  state: ICommandOutputState = {},
-  action: any
-) {
+export default function commandOutput(state = {}, action) {
   switch (action.type) {
     case EXECUTE_COMMAND_COMPLETE: {
       const { command, result, device } = action;
