+++ src/reducers/currentDevice.js
@@ -1,6 +1,6 @@
-export default function(state = "", action: any) {
+export default function(state = '', action) {
   switch (action.type) {
-    case "SELECT_DEVICE":
+    case 'SELECT_DEVICE':
       return action.name;
     default:
       return state;
