+++ src/reducers/rootReducer.js
@@ -1,0 +1,33 @@
+import { combineReducers } from 'redux';
+
+import deviceDetail from './deviceDetail';
+import deviceList from './deviceList';
+import loadingStatus from './loadingStatus';
+import commandOutput from './commandOutput';
+
+import currentDevice from './currentDevice';
+import devices from './devices';
+import attributes from './attributes';
+import commands from './commands';
+import properties from './properties';
+import modal from './modals';
+import user from './user';
+
+import error from './error';
+
+const rootReducer = combineReducers({
+  deviceDetail,
+  deviceList,
+  devices,
+  loadingStatus,
+  commandOutput,
+  currentDevice,
+  attributes,
+  commands,
+  properties,
+  error,
+  modal,
+  user
+});
+
+export default rootReducer;
