+++ src/selectors/commandOutput.js
@@ -1,0 +1,3 @@
+export default function getCommandOutputState(state) {
+  return state.commandOutput;
+}
