--- src/selectors/commandOutput.ts
@@ -1,5 +1,0 @@
-import { IRootState } from '../reducers/rootReducer';
-
-export function getCommandOutputState(state: IRootState) {
-    return state.commandOutput;
-}
