+++ src/selectors/error.js
@@ -1,0 +1,7 @@
+export function getError(state) {
+  return state.error;
+}
+
+export function hasError(state) {
+  return state.error && state.error !== '';
+}
