--- src/selectors/error.ts
@@ -1,9 +1,0 @@
-import { IRootState } from '../reducers/rootReducer';
-
-export function getError(state: IRootState) {
-    return state.error;
-}
-
-export function hasError(state: IRootState){
-    return state.error && state.error !== "";
-}
