+++ src/utils.js
@@ -1,7 +1,7 @@
-export function unique<T>(arr: T[]) {
+export function unique(arr) {
   return arr.filter((x, i) => arr.indexOf(x) === i);
 }
 
-export function objectValues<T>(obj: {[key: string]: T}) {
+export function objectValues(obj) {
   return obj ? Object.keys(obj).map(key => obj[key]) : [];
 }
